requirements.txt
openai
Twilio
Flask
python-dotenv
#-e .